an enum and using tags with playersettingarray maybe bebest. i run the foreach regularly is it more efficient just at the base hierachy to give a script it playersettings there once and everything just finds it from their basement

+++ Assets/Scripts/getDaysInput.cs
@@ -17,7 +17,7 @@ public class getDaysInput : MonoBehaviour {
 
 	public void runGetDaysInput(){
 //		Debug.Log("changing number of days from " + GameManager.numberDays + " game time was  " + GameManager.gameLength);
-	GameManager.numberDays = int.Parse( GetComponentInChildren<InputField>().text); // again would be better static and then triggered on button press
+	GameManager.numberDays = float.Parse( GetComponentInChildren<InputField>().text); // again would be better static and then triggered on button press
 //		Debug.Log("to " + GameManager.numberDays + " game time is  " + GameManager.gameLength);
 	}
 }
